# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is i9 GenieACS Proxy for API documents",
        "title": "Domain Nat Core Service",
        "termsOfService": "https://praction.in/terms/",
        "contact": {
            "name": "I9 API Support",
            "url": "http://www.praction.in/support",
            "email": "support@praction.in"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/acs-proxy/log-level": {
            "post": {
                "description": "With this endpoint you can setup log level for domain-service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Setup log level for domain-service",
                "parameters": [
                    {
                        "description": "Log Message",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Logging"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/acs-proxy/tasks/{task_id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a scheduled or completed ACS task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete Task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeviceResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    }
                }
            }
        },
        "/acs-proxy/tasks/{task_id}/retry": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retry a failed or pending ACS task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Retry Task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeviceResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    }
                }
            }
        },
...
        "/swagger/json": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Serve the OpenAPI documentation in JSON format",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Swagger"
                ],
                "summary": "Get Swagger JSON documentation",
                "responses": {
                    "200": {
                        "description": "Swagger JSON content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Swagger JSON file not found",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - Failed to read or serve the file",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    }
                }
            }
        },
        "/swagger/yaml": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Serve the OpenAPI documentation in YAML format",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Swagger"
                ],
                "summary": "Get Swagger YAML documentation",
                "responses": {
                    "200": {
                        "description": "Swagger YAML content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Swagger YAML file not found",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - Failed to read or serve the file",
                        "schema": {
                            "$ref": "#/definitions/models.BaseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.BaseError": {
            "description": "Generic error response wrapper.",
            "type": "object",
            "properties": {
                "errors": {
                    "description": "Errors is a list of detailed field-level errors.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ErrorDetail"
                    }
                },
                "message": {
                    "description": "Message is a human-readable error message.\nExample: Invalid input",
                    "type": "string",
                    "example": "Invalid input"
                },
                "status": {
                    "description": "Status of the response (e.g., error).\nExample: error",
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "description": "StatusCode is the HTTP status code.\nExample: 400",
                    "type": "integer",
                    "example": 400
                }
            }
        },
...
        "models.DeviceID": {
            "type": "object",
            "properties": {
                "_Manufacturer": {
                    "type": "string"
                },
                "_OUI": {
                    "type": "string"
                },
                "_ProductClass": {
                    "type": "string"
                },
                "_SerialNumber": {
                    "type": "string"
                }
            }
        },
        "models.DeviceModel": {
            "type": "object",
            "properties": {
                "FactoryReset": {
                    "type": "object",
                    "additionalProperties": true
                },
                "InternetGatewayDevice": {
                    "type": "object",
                    "additionalProperties": true
                },
                "Reboot": {
                    "type": "object",
                    "additionalProperties": true
                },
                "VirtualParameters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "_deviceId": {
                    "$ref": "#/definitions/models.DeviceID"
                },
                "_id": {
                    "type": "string"
                },
                "_lastBoot": {
                    "type": "string"
                },
                "_lastInform": {
                    "type": "string"
                },
                "_registered": {
                    "type": "string"
                },
                "_timestamp": {
                    "type": "string"
                }
            }
        },
        "models.DeviceResponseModel": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data contains pagination and list of app messengers",
                    "type": "object",
                    "properties": {
                        "data": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DeviceModel"
                            }
                        },
                        "pagination": {
                            "$ref": "#/definitions/models.MetaMode"
                        }
                    }
                },
                "message": {
                    "description": "Descriptive message\nExample: Configs fetched successfully",
                    "type": "string",
                    "example": "Configs fetched successfully"
                },
                "status": {
                    "description": "Status of the API call\nExample: success",
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "description": "HTTP status code\nExample: 200",
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "models.ErrorDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "Field that caused the error.\nExample: email",
                    "type": "string",
                    "example": "email"
                },
                "message": {
                    "description": "Message describing the error.\nExample: Email is required",
                    "type": "string",
                    "example": "Email is required"
                }
            }
        },
...
        "models.Logging": {
            "type": "object",
            "required": [
                "logLevel"
            ],
            "properties": {
                "logLevel": {
                    "type": "string",
                    "enum": [
                        "trace",
                        "debug",
                        "info",
                        "warn",
                        "error"
                    ]
                }
            }
        },
        "models.MetaMode": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Limit applied to this query.\nExample: 10",
                    "type": "integer",
                    "example": 10
                },
                "offset": {
                    "description": "Offset/Skip used in this query.\nExample: 0",
                    "type": "integer",
                    "example": 0
                },
                "total": {
                    "description": "Total number of matching documents.\nExample: 100",
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and then your key",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "description": "Authentication via cookie. Set by the server after successful login.",
            "type": "apiKey",
            "name": "auth",
            "in": "cookie"
        }
    },
...
}